+++ src/service/payment.ts
@@ -6,9 +6,12 @@ import { MongoRepository as EventRepo } from '../repo/event';
 import { MongoRepository as ProjectRepo } from '../repo/project';
 import { MongoRepository as SellerRepo } from '../repo/seller';
 import { MongoRepository as TaskRepo } from '../repo/task';
+import { MongoRepository as TransactionRepo } from '../repo/transaction';
+import { RedisRepository as TransactionNumberRepo } from '../repo/transactionNumber';
 
 import * as factory from '../factory';
 
+import * as AccountPaymentService from './payment/account';
 import * as CreditCardPaymentService from './payment/creditCard';
 import * as MovieTicketPaymentService from './payment/movieTicket';
 
@@ -26,6 +29,10 @@ export function pay(params: factory.task.pay.IData) {
         const paymentServiceType = params.object[0]?.typeOf;
 
         switch (paymentServiceType) {
+            case factory.service.paymentService.PaymentServiceType.Account:
+                await AccountPaymentService.payAccount(params)(repos);
+                break;
+
             case factory.service.paymentService.PaymentServiceType.CreditCard:
                 await CreditCardPaymentService.payCreditCard(params)(repos);
                 break;
@@ -48,6 +55,10 @@ export function voidPayment(params: factory.task.voidPayment.IData) {
         const paymentServiceType = params.object.object.typeOf;
 
         switch (paymentServiceType) {
+            case factory.service.paymentService.PaymentServiceType.Account:
+                await AccountPaymentService.voidTransaction(params)(repos);
+                break;
+
             case factory.service.paymentService.PaymentServiceType.CreditCard:
                 await CreditCardPaymentService.voidTransaction(params)(repos);
                 break;
@@ -71,11 +82,16 @@ export function refund(params: factory.task.refund.IData) {
         // event: EventRepo;
         project: ProjectRepo;
         seller: SellerRepo;
-        // task: TaskRepo;
+        transaction: TransactionRepo;
+        transactionNumber: TransactionNumberRepo;
     }) => {
         const paymentServiceType = params.object[0]?.typeOf;
 
         switch (paymentServiceType) {
+            case factory.service.paymentService.PaymentServiceType.Account:
+                await AccountPaymentService.refundAccount(params)(repos);
+                break;
+
             case factory.service.paymentService.PaymentServiceType.CreditCard:
                 await CreditCardPaymentService.refundCreditCard(params)(repos);
                 break;